+++ src/main/java/propertyFiles/Bundle.properties
@@ -2571,6 +2571,8 @@ pids.datacite.errors.DoiOnly=Only doi: is supported.
 
 #PublishDatasetCommand
 publishDatasetCommand.pidNotReserved=Cannot publish dataset because its persistent identifier has not been reserved.
+publishDatasetCommand.pidNotReserved=Cannot publish dataset because a persistent identifier for one or more of its files has not been reserved.
+
 
 # APIs
 api.errors.invalidApiToken=Invalid API token.
